requirements.txt

langchain==0.3.27
langchain-community==0.3.29
langchain-core==0.3.75
langchain-google-genai==2.1.10
python-dotenv==1.1.1
requests==2.32.5
beautifulsoup4==4.13.5

